+++ src/Battleship/Placement/RandomShipPlacer.cs
@@ -12,7 +12,7 @@ namespace Battleship.Placement
         private readonly Random rng = new();
         private const int MAX_TRIES_TO_PLACE_EACH_SHIP = 50;
 
-        public bool TryPlaceShips(GridState grid, params uint[] shipLengths)
+        public bool TryPlaceShips(GridState grid, params int[] shipLengths)
         {
             var sortedLengths = shipLengths.OrderByDescending(l => l);
 
@@ -22,8 +22,8 @@ namespace Battleship.Placement
                 var i = 0;
                 do
                 {
-                    var x = (uint)rng.Next(0, (int)grid.Width - 1);
-                    var y = (uint)rng.Next(0, (int)grid.Height - 1);
+                    var x = rng.Next(0, grid.Width - 1);
+                    var y = rng.Next(0, grid.Height - 1);
                     var orientation = rng.Next(0, 2) == 0 ? Orientation.Horizontal : Orientation.Vertical;
 
                     shipLocation = grid.TryPlaceShip(x, y, length, orientation);
