+++ src/Battleship/ShipLocation.cs
@@ -10,10 +10,10 @@ namespace Battleship
 {
     public class ShipLocation
     {
-        public uint X { get; private set; }
-        public uint Y { get; private set; }
+        public int X { get; private set; }
+        public int Y { get; private set; }
         public Orientation Orientation { get; private set; }
-        public uint Length { get; private set; }
+        public int Length { get; private set; }
 
         /// <summary>
         /// Lists the coordinates of each square that contains part of this ship.
@@ -22,15 +22,15 @@ namespace Battleship
         {
             get
             {
-                return Enumerable.Range(0, (int)Length).Select(i => {
+                return Enumerable.Range(0, Length).Select(i => {
                     var x = X + (Orientation == Orientation.Horizontal ? i : 0);
                     var y = Y + (Orientation == Orientation.Vertical ? i : 0);
-                    return new Point((int)x, (int)y);
+                    return new Point(x, y);
                     });
             }
         }
 
-        public ShipLocation(uint x, uint y, uint length, Orientation orientation)
+        public ShipLocation(int x, int y, int length, Orientation orientation)
         {
             X = x;
             Y = y;
