+++ test/Battleship.Test/ShipLocationTests.cs
@@ -14,12 +14,12 @@ namespace Battleship.Test
         [Fact]
         public void FullCoordinates_Horizontal()
         {
-            uint length = 4;
+            int length = 4;
             var ship = new ShipLocation(1, 1, length, Orientation.Horizontal);
 
             var result = ship.FullCoordinates;
 
-            Assert.Equal((int)length, result.Count());
+            Assert.Equal(length, result.Count());
             Assert.Contains(new Point(1, 1), result);
             Assert.Contains(new Point(2, 1), result);
             Assert.Contains(new Point(3, 1), result);
@@ -29,12 +29,12 @@ namespace Battleship.Test
         [Fact]
         public void FullCoordinates_Vertical()
         {
-            uint length = 4;
+            int length = 4;
             var ship = new ShipLocation(6, 1, length, Orientation.Vertical);
 
             var result = ship.FullCoordinates;
 
-            Assert.Equal((int)length, result.Count());
+            Assert.Equal(length, result.Count());
             Assert.Contains(new Point(6, 1), result);
             Assert.Contains(new Point(6, 2), result);
             Assert.Contains(new Point(6, 3), result);
